requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
jinja2==3.1.4
python-multipart==0.0.12
pydantic-settings==2.5.2
celery[redis]==5.4.0
ollama==0.6.1
